+++ src/lib.rs
@@ -8,7 +8,9 @@ mod http_languages;
 mod http_process;
 mod ip_options;
 mod mtu;
+mod observable_http_signals_matching;
 mod observable_signals;
+mod observable_tcp_signals_matching;
 pub mod p0f_output;
 mod process;
 mod signature_matcher;
